Removed ebuilds available in main portage tree
These ebuilds were originally added to allow for unstable versions to be 
installed through the nitelite overlay. This functionality has been moved to the 
puppet-nitelite repository and the ebuilds, therefore, no longer need to be 
managed in this overlay.

This overlay will now contain only custom ebuilds that are not available in the 
main portage tree.